Handle Users.messages.get "Not Found" errors more explicitly.
Before any HTTP 404 would be treated as a "message not found"
condition.  We now insist that the Gmail API returns a structured
"notFound" error in the response body's JSON.  This increases the
chance that the error is returned by the Gmail backend and not
something else such as a proxy sitting between the client and Gmail.